+++ src/modeljacobian.cpp
@@ -148,7 +148,7 @@ bool ModelJacobian::computeFast(int nrhs, const mxArray* prhs[])
 
 bool ModelJacobian::processArguments(int nrhs, const mxArray * prhs[])
 {
-  int numDof = modelState->dof();
+  size_t numDof = modelState->dof();
 
   if(mxGetM(prhs[1]) != 9 || mxGetN(prhs[1]) != 1 || mxGetM(prhs[2]) != 3 || mxGetN(prhs[2]) != 1 || mxGetM(prhs[3]) != numDof || mxGetN(prhs[3]) != 1 || !mxIsChar(prhs[4]))
   {
@@ -178,7 +178,7 @@ bool ModelJacobian::processArguments(int nrhs, const mxArray * prhs[])
 #ifdef DEBUG
   mexPrintf("qj received \n");
 
-  for(int i = 0; i< numDof;i++)
+  for(size_t i = 0; i< numDof;i++)
   {
     mexPrintf(" %f",qj[i]);
   }
@@ -205,7 +205,7 @@ bool ModelJacobian::processArguments(int nrhs, const mxArray * prhs[])
     }
   }
   int columnMajCtr = 0;
-  for (int i = 0; i<(numDof+6); i++)
+  for (size_t i = 0; i<(numDof+6); i++)
   {
       for (int j = 0;j<6;j++)
       {
